Move readline wrapping to a separate module
